Split heap_fold_call_name into a pattern-1 name router; split mod_p4_c.rs to clear max-lines
